third_party/blink/web_tests/http/tests/navigation/anchor-frames-gbk-expected.txt



--------
Frame: 'main'
--------
Tests that loading a frame with a URL that contains a fragment pointed at a named anchor actually scrolls to that anchor.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS document.body.offsetHeight > document.documentElement.clientHeight is true
FAIL document.documentElement.scrollTop > 0 should be true. Was false.
FAIL document.documentElement.scrollTop + document.documentElement.clientHeight > 2000 should be true. Was false.
PASS successfullyParsed is true

TEST COMPLETE

This is an anchor point named as the Unicode equivalent of the GBK sequence %a9g (test trailing low byte).

--------
Frame: 'footer'
--------
